--- src/main/kotlin/racingcar/PrintUnits.kt
@@ -1,11 +1,0 @@
-package racingcar
-
-object PrintUtils {
-    fun printIntroduction() {
-        println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)")
-    }
-
-    fun printTryCount() {
-        println("시도할 횟수는 몇 회인가요?")
-    }
-}
