+++ src/main/kotlin/racingcar/RacingGame.kt
@@ -3,7 +3,7 @@ package racingcar
 class RacingGame(private val cars: List<Car>, private val tryCount: Int) {
 
     fun playGame() {
-        println("\n실행 결과")
+        PrintUtils.printGameIntroduction()
         repeat(tryCount) {
             playRound()
         }
@@ -16,14 +16,13 @@ class RacingGame(private val cars: List<Car>, private val tryCount: Int) {
             car.move()
             roundResults.add(car.toString())
         }
-        println(roundResults.joinToString("\n"))
-        println()
+        PrintUtils.printRoundResults(roundResults)
     }
 
     fun printWinners() {
         val maxPosition = cars.maxOf { it.position }
         val winners = cars.filter { it.position == maxPosition }
         val winnerNames = winners.joinToString(", ") { it.name }
-        println("최종 우승자: $winnerNames")
+        PrintUtils.printWinners(winnerNames)
     }
 }
