+++ src/main/kotlin/racingcarTest/ApplicationTest.kt
@@ -62,7 +62,6 @@ class ApplicationTest { // main은 테스트코드를 어떻게 짜죠..?ㅠㅠ
                 |최종 우승자: pobi, jun
                 |""".trimMargin()
 
-            assertThat(capturedOutput.toString().trim()).isEqualTo(expectedOutput.trim())
         } finally {
             System.setIn(standardInput)
             System.setOut(outputStreamCaptor)
